requirements.txt
instagrapi==2.1.3
python-dotenv==1.0.1